+++ cricinfo/source_models/team.py
@@ -1,6 +1,8 @@
 from abc import ABC
 
-from cricinfo.source_models.common import CCBaseModel, NameMixin, RefMixin
+from pydantic import AliasChoices, Field
+
+from cricinfo.source_models.common import CCBaseModel, Event, NameMixin, RefMixin
 
 
 class TeamCommon(CCBaseModel, ABC):
@@ -15,3 +17,14 @@ class TeamWithName(TeamCommon, NameMixin): ...
 class TeamWithColorAndLogos(TeamCommon):
     color: str
     logos: list[RefMixin]
+
+
+class TeamFull(TeamWithName):
+    color: str
+    nickname: str
+    short_display_name: str
+    is_national: bool
+    is_active: bool
+    classes: list[int]
+    current_match: Event = Field(default=None, validation_alias=AliasChoices("event"))
+    current_players_link: RefMixin = Field(default=None, validation_alias=AliasChoices("athletes"))